arm64: KVM: Disable virtual timer even if the guest is not using it
[ Upstream commit c4cbba9fa078f55d9f6d081dbb4aec7cf969e7c7 ]

When running a guest with the architected timer disabled (with QEMU and
the kernel_irqchip=off option, for example), it is important to make
sure the timer gets turned off. Otherwise, the guest may try to
enable it anyway, leading to a screaming HW interrupt.

The fix is to unconditionally turn off the virtual timer on guest
exit.

Cc: stable@vger.kernel.org
Reviewed-by: Christoffer Dall <christoffer.dall@linaro.org>
Signed-off-by: Marc Zyngier <marc.zyngier@arm.com>
Signed-off-by: Sasha Levin <sasha.levin@oracle.com>
(cherry picked from commit 325f791748d7fab63f859c518f84f8320d835241)